+++ src/client/publish.ts
@@ -58,7 +58,7 @@ const getEventEnvelopeFromPublishedEvent = (event: PublishEvent): EventEnvelope 
             correlationId: event.correlationId ?? "",
             causationId: event.causationId ?? "",
             payload,
-            raisedAt: "0",
+            raisedAt: 0,
         },
     };
 };